Improve Dashboard component
