+++ packages/frappe-ui-react/src/components/dashboard/index.ts
@@ -1,4 +1,4 @@
-export { Dashboard, default } from "./dashboard";
+export { default as Dashboard } from "./dashboard";
 export { WidgetWrapper } from "./widgetWrapper";
 export { DashboardWidgetGallery } from "./dashboardWidgetGallery";
 export { DashboardProvider } from "./dashboardContext";
